+++ conf/gfpeps/default.yaml
@@ -1,6 +1,6 @@
 params:
-  Nv: 2 # Number of virtual fermions on each bond
-  seed: 42 # Random seed
+  Nv: 1 # Number of virtual fermions on each bond
+  seed: 44 # Random seed
 
 lattice:
   Lx: 101 # System size in x-direction
@@ -8,18 +8,19 @@ lattice:
 
 hamiltonian:
   ht: 1.0 # Hoping amplitude
-  DeltaX: 1.0 # d-wave paring amplitude in x-direction
-  DeltaY: -1.0 # d-wave paring amplitude in y-direction
+  DeltaX: 0.0 # d-wave paring amplitude in x-direction
+  DeltaY: 0.0 # d-wave paring amplitude in y-direction
   delta: 0.0 # desity of holes
   Mu: 0.0 # chemical potential
   solve_mu_from_delta: false # whether to solve chemical potential from hole density
 
 file:
-  LoadFile: "./data/default.h5" # Load initial T from this file, if it exists
+  # LoadFile: "./data/default.h5" # Load initial T from this file, if it exists
+  LoadFile: ""
   WriteFile: "./data/default.h5" # Write information about final Gaussian fPEPS to this file
 
 optimizer:
   MaxIter: 100
   gtol: 1E-7 # gtol for optimizer
 
-backend: gpu
+backend: cpu
