+++ src/gfpeps/__init__.py
@@ -2,4 +2,5 @@
 #
 # SPDX-License-Identifier: MIT
 
-from main import gaussian_fpeps
+__all__ = ['gaussian_fpeps']
+from .gaussian_fpeps import gaussian_fpeps
