+++ src/gfpeps/deltatomu.py
@@ -2,7 +2,7 @@
 import numpy as np
 import jax.numpy as jnp
 import jax
-from jax.config import config
+from jax import config
 config.update("jax_enable_x64", True)
 from scipy.optimize import root_scalar
