+++ src/gfpeps/loadwrite.py
@@ -1,24 +1,25 @@
 import numpy as np
 import jax.numpy as jnp
 import h5py,os
+import logging
 
 def initialT(loadfile,Tsize):
     if loadfile != None and os.path.isfile(loadfile):
-        print('Try to initialize T from',loadfile)
+        logging.info(f'Try to initialize T from {loadfile}')
         with h5py.File(loadfile, 'r') as f:
             if "/transformer/T" in f.keys():
                 T = f["/transformer/T"][:]
                 f.close()
                 return jnp.reshape(T,(Tsize,Tsize))
             else:
-                print("Load Failed! No /transformer/T in",loadfile," switch to random initialize!")
+                logging.debug(f"Load Failed! No /transformer/T in {loadfile} switch to random initialize!")
                 f.close()
                 return jnp.array(np.random.rand(Tsize,Tsize))
     return jnp.array(np.random.rand(Tsize,Tsize))
 
 def savelog(writefile,results):
     if writefile != None:
-        print('Save T to',writefile)
+        logging.info(f'Save T to {writefile}')
         with h5py.File(writefile, 'w') as f:
             f["/transformer/T"] = results.x
             f["/energy/EABD"] = results.fun
@@ -30,7 +31,7 @@ def savelog(writefile,results):
 
 def savelog_trivial(writefile,x,fun,Eg,args):
     if writefile != None:
-        print('Save T to',writefile)
+        logging.info(f'Save T to {writefile}')
         with h5py.File(writefile, 'w') as f:
             f["/transformer/T"] = x
             f["/energy/EABD"] = fun
